+++ PopditMobile/Models/BubbleMobile.cs
@@ -3,8 +3,8 @@
     public partial class BubbleMobile
     { 
         public int Id { get; set; }
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+        public double Latitude { get; set; }
+        public double Longitude { get; set; }
         public int Radius { get; set; }
     }
 }
